requirements.txt
flask
six
python-ldap
PyJWT
flask-restplus
apispec
webargs
flask-marshmallow
# flask-cors
flatdict
bidict
bson
